+++ NEWS.md
@@ -5,6 +5,10 @@
   - use the presence of intronic read coverage to indicate `de-novo` transcription
   - apply first to well studied organisms (drosophila, zebrafish), then use as method for less well characterized systems, such as sea urchin. 
   
+## 2018-02-15 Make pipeline species agnostic 
+  - reorganize pipeline to handle additional species
+  - download zebrafish and xenopus datasets from GEO
+  
 ## 2018-02-14 Finish drosophila characterization
   - Pick sensible cutoffs for zygotic and maternal
   - Plot as heatmaps, ordering by time to first expression for strictly zygotic